backend/requirements.txt
flask
flask-cors
deepface
opencv-python
pandas
numpy
tf-keras
kagglehub
hsemotion
transformers
Pillow
